requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
google-generativeai
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dotenv==0.9.9
elastic-transport==9.1.0
elasticsearch==9.1.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-pydantic==0.6
jsonschema-specifications==2025.4.1
langchain==0.3.27

langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
mcp==1.13.0
mcp-use==1.3.10
multidict==6.6.4
openai==1.99.9
orjson==3.11.2
packaging==25.0
posthog==6.5.0
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
scarf-sdk==0.1.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.24.0
